api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
google-cloud-firestore
google-cloud-dialogflow
google-cloud-bigquery
google-cloud-aiplatform
pydantic==2.7.1
websockets==12.0
requests
